FS-browser: pass OIDC token to Azure Block Download
